Animations
Slay has an idle animation that faces left which is also used when he is move left. But I have to make it so that when the D key is pressed, that he is move right and facing towards right. Same thing applies to Saquon but not in the game yet.